+++ examples/coroutinesExample/src/main/java/io/realm/examples/coroutinesexample/model/Dog.kt
@@ -1,0 +1,11 @@
+package io.realm.examples.coroutinesexample.model
+
+import io.realm.RealmObject
+import io.realm.annotations.PrimaryKey
+
+open class Dog : RealmObject() {
+    @PrimaryKey
+    var name: String = ""
+    var age: Int = 1
+    var owner: String = ""
+}