+++ CHANGELOG.md
@@ -17,6 +17,10 @@ The old Realm Cloud legacy APIs have undergone significant refactoring. The new 
 * Users can now opt out from allowing queries to be launched from the UI thread by using `RealmConfiguration.Builder.allowQueriesOnUiThread(false)`. A `RealmException` will be thrown when calling `RealmQuery.findAll`, `RealmQuery.findFirst`, `RealmQuery.minimumDate`, `RealmQuery.maximumDate`, `RealmQuery.count`, `RealmQuery.sum`, `RealmQuery.max`, `RealmQuery.min`, `RealmQuery.average` and `RealmQuery.averageDecimal128` from the UI thread after having used `allowQueriesOnUiThread(false)`. Queries will be allowed from the thread from which the Realm instance was obtained as it always has been by default, although we recommend using a non-UI thread to launch them, or, alternatively, using `RealmQuery.findAllAsync` or `RealmQuery.findFirstAsync`.
 * `BaseRealm.refresh` will throw a `RealmException` if it is being called from the UI thread if `allowQueriesOnUiThread` is set to `false`, though it will be allowed by default.
 * Added `DynamicRealm.executeTransactionAsync` as a by-product of the new `allowWritesOnUiThread` configuration setting.
+* Added Kotlin extension suspend function `Realm.executeTransactionAwait` which runs transactions inside coroutines.
+* Added Kotlin extension function `RealmResults.toFlow` which returns a Kotlin flow, similar to RxJava's convenience method `asFlowable`.
+* Added Kotlin extension function `RealmList.toFlow` which returns a Kotlin flow, similar to RxJava's convenience method `asFlowable`.
+* Added Kotlin extension function `RealmModel.toFlow` which returns a Kotlin flow, similar to RxJava's convenience method `asFlowable`.
 
 
 ## 10.0.0-BETA.8 (2020-09-23)